+++ packages/expo-ui/src/swift-ui/TextField/index.tsx
@@ -146,3 +146,6 @@ export function TextField(props: TextFieldProps) {
 }
 
 TextField.Placeholder = Placeholder;
+
+// Exported for docs api data
+export { type ObservableState };